Moved oauth2 functions to own file
